move `sharedProcessLifecycleService`
