+++ src/vs/platform/lifecycle/node/sharedProcessLifecycleService.ts
